IGNITE-3670 IGFS: Local secondary: Add test for listPath() method called for symlink to directory - fix review issues
