+++ modules/core/src/test/java/org/apache/ignite/internal/processors/igfs/IgfsLocalSecondaryFileSystemDualAbstractSelfTest.java
@@ -17,12 +17,21 @@
 
 package org.apache.ignite.internal.processors.igfs;
 
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.file.Files;
+import java.util.Collection;
+import org.apache.ignite.igfs.IgfsFile;
 import org.apache.ignite.igfs.IgfsMode;
+import org.apache.ignite.igfs.IgfsPath;
 import org.apache.ignite.igfs.secondary.IgfsSecondaryFileSystem;
 import org.apache.ignite.igfs.secondary.local.LocalIgfsSecondaryFileSystem;
+import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.internal.U;
 
 import java.io.File;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Abstract test for Hadoop 1.0 file system stack.
@@ -32,12 +41,43 @@ public abstract class IgfsLocalSecondaryFileSystemDualAbstractSelfTest extends I
     private static final String FS_WORK_DIR = U.getIgniteHome() + File.separatorChar + "work"
         + File.separatorChar + "fs";
 
+    /** */
+    private static final String FS_EXT_DIR = U.getIgniteHome() + File.separatorChar + "work"
+        + File.separatorChar + "ext";
+
+    /** */
+    private final File dirLinkDest = new File(FS_EXT_DIR + File.separatorChar + "extdir");
+
+    /** */
+    private final File fileLinkDest =
+        new File(FS_EXT_DIR + File.separatorChar + "extdir" + File.separatorChar + "filedest");
+
+    /** */
+    private final File dirLinkSrc = new File(FS_WORK_DIR + File.separatorChar + "dir");
+
+    /** */
+    private final File fileLinkSrc = new File(FS_WORK_DIR + File.separatorChar + "file");
+
+
     /** Constructor.
      * @param mode IGFS mode.
      */
     public IgfsLocalSecondaryFileSystemDualAbstractSelfTest(IgfsMode mode) {
         super(mode);
     }
+
+    /** {@inheritDoc} */
+    @Override protected void beforeTest() throws Exception {
+        super.beforeTest();
+
+        final File extDir = new File(FS_EXT_DIR);
+
+        if (!extDir.exists())
+            assert extDir.mkdirs();
+        else
+            cleanDirectory(extDir);
+    }
+
 
     /**
      * Creates secondary filesystems.
@@ -72,5 +112,99 @@ public abstract class IgfsLocalSecondaryFileSystemDualAbstractSelfTest extends I
     /** {@inheritDoc} */
     @Override protected boolean timesSupported() {
         return false;
+    }
+
+
+    /**
+     * @param dir Directory to clean.
+     */
+    private static void cleanDirectory(File dir){
+        File[] entries = dir.listFiles();
+
+        if (entries != null) {
+            for (File entry : entries) {
+                if (entry.isDirectory()) {
+                    cleanDirectory(entry);
+                    assert entry.delete();
+                }
+                else
+                    assert entry.delete();
+            }
+        }
+    }
+
+    /**
+     * @param f File object.
+     * @param overwrite Overwrite flag.
+     * @param chunks File content.
+     * @throws IOException If failed.
+     */
+    protected static void createFile(File f, boolean overwrite, @Nullable byte[]... chunks) throws IOException {
+        OutputStream os = null;
+
+        try {
+            os = new FileOutputStream(f, overwrite);
+
+            writeFileChunks(os, chunks);
+        }
+        finally {
+            U.closeQuiet(os);
+        }
+    }
+
+    /**
+     *
+     * @throws Exception If failed.
+     */
+    @SuppressWarnings("ConstantConditions")
+    public void testListPathForSymlink() throws Exception {
+        createSymlinks();
+
+        assertTrue(igfs.info(DIR).isDirectory());
+
+        Collection<IgfsPath> pathes = igfs.listPaths(DIR);
+        Collection<IgfsFile> files = igfs.listFiles(DIR);
+
+        assertEquals(1, pathes.size());
+        assertEquals(1, files.size());
+
+        assertEquals("filedest", F.first(pathes).name());
+        assertEquals("filedest", F.first(files).path().name());
+    }
+
+    /**
+     *
+     * @throws Exception If failed.
+     */
+    public void testDeleteSymlinkDir() throws Exception {
+        createSymlinks();
+
+        // Only symlink must be deleted. Destination content must be exist.
+        igfs.delete(DIR, true);
+
+        assertTrue(fileLinkDest.exists());
+    }
+
+    /**
+     *
+     * @throws Exception If failed.
+     */
+    public void testSymlinkToFile() throws Exception {
+        createSymlinks();
+
+        checkFileContent(igfs, new IgfsPath("/file"), chunk);
+    }
+
+    /**
+     *
+     * @throws Exception If failed.
+     */
+    private void createSymlinks() throws Exception {
+        assert dirLinkDest.mkdir();
+
+        createFile(fileLinkDest, true, chunk);
+
+        Files.createSymbolicLink(dirLinkSrc.toPath(), dirLinkDest.toPath());
+        Files.createSymbolicLink(fileLinkSrc.toPath(), fileLinkDest.toPath());
     }
 }
